+++ lib/generators/statesman/generator_helpers.rb
@@ -28,9 +28,17 @@ module Statesman
       klass.demodulize.underscore.pluralize
     end
 
+    def index_name(index_id)
+      "index_#{table_name}_#{index_id}"
+    end
+
     def mysql?
-      ActiveRecord::Base.configurations[Rails.env]
-        .try(:[], "adapter").try(:match, /mysql/)
+      ActiveRecord::Base.configurations[Rails.env].
+        try(:[], "adapter").try(:match, /mysql/)
+    end
+
+    def database_supports_partial_indexes?
+      Statesman::Adapters::ActiveRecord.database_supports_partial_indexes?
     end
   end
 end