+++ lib/statesman/machine.rb
@@ -77,7 +77,7 @@ module Statesman
         if event
           events[event] ||= {}
           events[event][from] ||= []
-          events[event][from]  += to
+          events[event][from] += to
         end
       end
 
@@ -183,7 +183,7 @@ module Statesman
       @object = object
       @transition_class = options[:transition_class]
       @storage_adapter = adapter_class(@transition_class).new(
-        @transition_class, object, self)
+        @transition_class, object, self, options)
       send(:after_initialize) if respond_to? :after_initialize
     end