+++ spec/generators/statesman/active_record_transition_generator_spec.rb
@@ -12,7 +12,7 @@ describe Statesman::ActiveRecordTransitionGenerator, type: :generator do
     subject { file('app/models/yummy/bacon_transition.rb') }
 
     it { is_expected.to contain(/:bacon_transition/) }
-    it { is_expected.not_to contain(/:yummy\/bacon/) }
+    it { is_expected.not_to contain(%r{:yummy/bacon}) }
     it { is_expected.to contain(/class_name: 'Yummy::Bacon'/) }
   end