+++ spec/generators/statesman/migration_generator_spec.rb
@@ -26,7 +26,17 @@ describe Statesman::MigrationGenerator, type: :generator do
     end
 
     it { is_expected.to contain(/:bacon_transition/) }
-    it { is_expected.not_to contain(/:yummy\/bacon/) }
+    it { is_expected.not_to contain(%r{:yummy/bacon}) }
     it { is_expected.to contain(/null: false/) }
+
+    it "names the sorting index appropriately" do
+      expect(subject).
+        to contain("name: \"index_bacon_transitions_parent_sort\"")
+    end
+
+    it "names the most_recent index appropriately" do
+      expect(subject).
+        to contain("name: \"index_bacon_transitions_parent_most_recent\"")
+    end
   end
 end