+++ spec/spec_helper.rb
@@ -1,9 +1,15 @@
 require "statesman"
 require "sqlite3"
 require "active_record"
+# We have to include all of Rails to make rspec-rails work
+require "rails"
+require "action_view"
+require "action_dispatch"
+require "action_controller"
+require "rspec/rails"
 require "support/active_record"
 require "mongoid"
-require 'rspec/its'
+require "rspec/its"
 
 RSpec.configure do |config|
   config.raise_errors_for_deprecations!
@@ -26,12 +32,28 @@ RSpec.configure do |config|
     raise(error)
   end
 
-  config.before(:each) do
-    # Connect to & cleanup test database
-    ActiveRecord::Base.establish_connection(adapter: "sqlite3",
-                                            database: DB.to_s)
+  if config.exclusion_filter[:active_record]
+    puts "Skipping ActiveRecord tests"
+  else
+    # Connect to the database for activerecord tests
+    db_conn_spec = ENV["DATABASE_URL"]
+    db_conn_spec ||= { adapter: "sqlite3", database: ":memory:" }
+    ActiveRecord::Base.establish_connection(db_conn_spec)
 
-    %w(my_models my_model_transitions).each do |table_name|
+    db_adapter = ActiveRecord::Base.connection.adapter_name
+    puts "Running with database adapter '#{db_adapter}'"
+  end
+
+  config.before(:each, active_record: true) do
+    tables = %w(
+      my_active_record_models
+      my_active_record_model_transitions
+      my_namespace_my_active_record_models
+      my_namespace_my_active_record_model_transitions
+      other_active_record_models
+      other_active_record_model_transitions
+    )
+    tables.each do |table_name|
       sql = "DROP TABLE IF EXISTS #{table_name};"
       ActiveRecord::Base.connection.execute(sql)
     end
@@ -45,9 +67,30 @@ RSpec.configure do |config|
     def prepare_transitions_table
       silence_stream(STDOUT) do
         CreateMyActiveRecordModelTransitionMigration.migrate(:up)
+        MyActiveRecordModelTransition.reset_column_information
       end
     end
-  end
 
-  config.after(:each) { DB.delete if DB.exist? }
+    def drop_most_recent_column
+      silence_stream(STDOUT) do
+        DropMostRecentColumn.migrate(:up)
+        MyActiveRecordModelTransition.reset_column_information
+      end
+    end
+
+    def prepare_other_model_table
+      silence_stream(STDOUT) do
+        CreateOtherActiveRecordModelMigration.migrate(:up)
+      end
+    end
+
+    def prepare_other_transitions_table
+      silence_stream(STDOUT) do
+        CreateOtherActiveRecordModelTransitionMigration.migrate(:up)
+        OtherActiveRecordModelTransition.reset_column_information
+      end
+    end
+
+    MyNamespace::MyActiveRecordModelTransition.serialize(:metadata, JSON)
+  end
 end