+++ spec/statesman/adapters/active_record_queries_spec.rb
@@ -1,10 +1,17 @@
 require "spec_helper"
 
-describe Statesman::Adapters::ActiveRecordQueries do
+describe Statesman::Adapters::ActiveRecordQueries, active_record: true do
   before do
     prepare_model_table
     prepare_transitions_table
+    prepare_other_model_table
+    prepare_other_transitions_table
   end
+
+  before do
+    Statesman.configure { storage_adapter(Statesman::Adapters::ActiveRecord) }
+  end
+  after { Statesman.configure { storage_adapter(Statesman::Adapters::Memory) } }
 
   before do
     MyActiveRecordModel.send(:include, Statesman::Adapters::ActiveRecordQueries)
@@ -17,17 +24,31 @@ describe Statesman::Adapters::ActiveRecordQueries do
         :initial
       end
     end
+
+    OtherActiveRecordModel.send(:include,
+                                Statesman::Adapters::ActiveRecordQueries)
+    OtherActiveRecordModel.class_eval do
+      def self.transition_class
+        OtherActiveRecordModelTransition
+      end
+
+      def self.initial_state
+        :initial
+      end
+    end
   end
+  before { MyActiveRecordModel.send(:has_one, :other_active_record_model) }
+  before { OtherActiveRecordModel.send(:belongs_to, :my_active_record_model) }
 
   let!(:model) do
     model = MyActiveRecordModel.create
-    model.my_active_record_model_transitions.create(to_state: :succeeded)
+    model.state_machine.transition_to(:succeeded)
     model
   end
 
   let!(:other_model) do
     model = MyActiveRecordModel.create
-    model.my_active_record_model_transitions.create(to_state: :failed)
+    model.state_machine.transition_to(:failed)
     model
   end
 
@@ -35,45 +56,163 @@ describe Statesman::Adapters::ActiveRecordQueries do
 
   let!(:returned_to_initial_model) do
     model = MyActiveRecordModel.create
-    model.my_active_record_model_transitions.create(to_state: :failed)
-    model.my_active_record_model_transitions.create(to_state: :initial)
+    model.state_machine.transition_to(:failed)
+    model.state_machine.transition_to(:initial)
     model
   end
 
-  describe ".in_state" do
-    context "given a single state" do
-      subject { MyActiveRecordModel.in_state(:succeeded) }
+  context "with a most_recent column" do
+    describe ".in_state" do
+      context "given a single state" do
+        subject { MyActiveRecordModel.in_state(:succeeded) }
 
-      it { is_expected.to include model }
+        it { is_expected.to include model }
+        it { is_expected.not_to include other_model }
+        its(:to_sql) { is_expected.to include('.most_recent ') }
+      end
+
+      context "given multiple states" do
+        subject { MyActiveRecordModel.in_state(:succeeded, :failed) }
+
+        it { is_expected.to include model }
+        it { is_expected.to include other_model }
+      end
+
+      context "given the initial state" do
+        subject { MyActiveRecordModel.in_state(:initial) }
+
+        it { is_expected.to include initial_state_model }
+        it { is_expected.to include returned_to_initial_model }
+      end
+
+      context "given an array of states" do
+        subject { MyActiveRecordModel.in_state([:succeeded, :failed]) }
+
+        it { is_expected.to include model }
+        it { is_expected.to include other_model }
+      end
+
+      context "merging two queries" do
+        subject do
+          MyActiveRecordModel.in_state(:succeeded).
+            joins(:other_active_record_model).
+            merge(OtherActiveRecordModel.in_state(:initial))
+        end
+
+        it { is_expected.to be_empty }
+      end
     end
 
-    context "given multiple states" do
-      subject { MyActiveRecordModel.in_state(:succeeded, :failed) }
+    describe ".not_in_state" do
+      context "given a single state" do
+        subject { MyActiveRecordModel.not_in_state(:failed) }
+        it { is_expected.to include model }
+        it { is_expected.not_to include other_model }
+        its(:to_sql) { is_expected.to include('.most_recent ') }
+      end
 
-      it { is_expected.to include model }
-      it { is_expected.to include other_model }
-    end
+      context "given multiple states" do
+        subject { MyActiveRecordModel.not_in_state(:succeeded, :failed) }
+        it do
+          is_expected.to match_array([initial_state_model,
+                                      returned_to_initial_model])
+        end
+      end
 
-    context "given the initial state" do
-      subject { MyActiveRecordModel.in_state(:initial) }
-
-      it { is_expected.to include initial_state_model }
-      it { is_expected.to include returned_to_initial_model }
+      context "given an array of states" do
+        subject { MyActiveRecordModel.not_in_state([:succeeded, :failed]) }
+        it do
+          is_expected.to match_array([initial_state_model,
+                                      returned_to_initial_model])
+        end
+      end
     end
   end
 
-  describe ".not_in_state" do
-    context "given a single state" do
-      subject { MyActiveRecordModel.not_in_state(:failed) }
-      it { is_expected.to include model }
-      it { is_expected.not_to include other_model }
+  context "without a most_recent column" do
+    before { drop_most_recent_column }
+
+    describe ".in_state" do
+      context "given a single state" do
+        subject { MyActiveRecordModel.in_state(:succeeded) }
+
+        it { is_expected.to include model }
+        its(:to_sql) { is_expected.not_to include('.most_recent ') }
+      end
+
+      context "given multiple states" do
+        subject { MyActiveRecordModel.in_state(:succeeded, :failed) }
+
+        it { is_expected.to include model }
+        it { is_expected.to include other_model }
+      end
+
+      context "given the initial state" do
+        subject { MyActiveRecordModel.in_state(:initial) }
+
+        it { is_expected.to include initial_state_model }
+        it { is_expected.to include returned_to_initial_model }
+      end
+
+      context "given an array of states" do
+        subject { MyActiveRecordModel.in_state([:succeeded, :failed]) }
+
+        it { is_expected.to include model }
+        it { is_expected.to include other_model }
+      end
     end
 
-    context "given multiple states" do
-      subject { MyActiveRecordModel.not_in_state(:succeeded, :failed) }
-      it do
-        is_expected.to match_array([initial_state_model,
-                                    returned_to_initial_model])
+    describe ".not_in_state" do
+      context "given a single state" do
+        subject { MyActiveRecordModel.not_in_state(:failed) }
+        it { is_expected.to include model }
+        it { is_expected.not_to include other_model }
+        its(:to_sql) { is_expected.not_to include('.most_recent ') }
+      end
+
+      context "given multiple states" do
+        subject { MyActiveRecordModel.not_in_state(:succeeded, :failed) }
+        it do
+          is_expected.to match_array([initial_state_model,
+                                      returned_to_initial_model])
+        end
+      end
+
+      context "given an array of states" do
+        subject { MyActiveRecordModel.not_in_state([:succeeded, :failed]) }
+        it do
+          is_expected.to match_array([initial_state_model,
+                                      returned_to_initial_model])
+        end
+      end
+    end
+  end
+
+  context "with a transition name" do
+    before do
+      MyActiveRecordModel.send(:has_many,
+                               :custom_name,
+                               class_name: 'MyActiveRecordModelTransition')
+      MyActiveRecordModel.class_eval do
+        def self.transition_name
+          :custom_name
+        end
+      end
+    end
+
+    context "with a most_recent column" do
+      describe ".in_state" do
+        subject(:query) { MyActiveRecordModel.in_state(:succeeded) }
+        specify { expect { query }.to_not raise_error }
+      end
+    end
+
+    context "without a most_recent column" do
+      before { drop_most_recent_column }
+
+      describe ".in_state" do
+        subject(:query) { MyActiveRecordModel.in_state(:succeeded) }
+        specify { expect { query }.to_not raise_error }
       end
     end
   end