+++ spec/statesman/adapters/mongoid_spec.rb
@@ -5,22 +5,16 @@ require "support/mongoid"
 require "mongoid"
 
 describe Statesman::Adapters::Mongoid, mongo: true do
-  after do
-    Mongoid.purge!
-  end
-  let(:observer) do
-    result = double(Statesman::Machine)
-    allow(result).to receive(:execute)
-    result
-  end
+  after { Mongoid.purge! }
+  let(:observer) { double(Statesman::Machine, execute: nil) }
   let(:model) { MyMongoidModel.create(current_state: :pending) }
   it_behaves_like "an adapter", described_class, MyMongoidModelTransition
 
   describe "#initialize" do
     context "with unserialized metadata" do
       before do
-        allow_any_instance_of(described_class)
-          .to receive_messages(transition_class_hash_fields: [])
+        allow_any_instance_of(described_class).
+          to receive_messages(transition_class_hash_fields: [])
       end
 
       it "raises an exception if metadata is not serialized" do
@@ -38,14 +32,12 @@ describe Statesman::Adapters::Mongoid, mongo: true do
     end
 
     context "with a previously looked up transition" do
-      before do
-        adapter.create(:x, :y)
-        adapter.last
-      end
+      before { adapter.create(:x, :y) }
+      before { adapter.last }
 
       it "caches the transition" do
-        expect_any_instance_of(MyMongoidModel)
-          .to receive(:my_mongoid_model_transitions).never
+        expect_any_instance_of(MyMongoidModel).
+          to receive(:my_mongoid_model_transitions).never
         adapter.last
       end